% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mat.R
\name{as.mat}
\alias{as.mat}
\alias{as.mat.data.frame}
\alias{as.mat.default}
\alias{as.mat.matrix}
\alias{is.mat}
\title{Coerce to a \code{"mat"} Object}
\usage{
as.mat(x, ...)

\method{as.mat}{default}(x, ...)

\method{as.mat}{matrix}(x, ...)

\method{as.mat}{data.frame}(x, ...)

is.mat(x)
}
\arguments{
\item{x}{A matrix or data frame.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
Functions to check if an object is of class \code{"mat"}, or coerce it if
possible.
}

