% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{falses}
\alias{falses}
\alias{trues}
\title{Matrix/Array of Logical Values}
\usage{
falses(nrow = 1, ncol = 1, ...)

trues(nrow = 1, ncol = 1, ...)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{...}{Further dimensions of the array.}
}
\value{
A matrix or array of logical values.
}
\description{
Construct a matrix or multi-way array of logical values.
}
\examples{
falses(3)  # column vector of FALSEs
falses(2, 3)  # 2-by-3 matrix of FALSEs
falses(2, 3, 2)  # 2-by-3-by-2 array of FALSEs
fill(FALSE, 2, 2, 2)
}
\seealso{
\code{\link{fill}}, \code{\link{ones}}, \code{\link{zeros}}.
}

