% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{is.triu}
\alias{is.triu}
\title{Upper Triangular Matrix Test}
\usage{
is.triu(x)
}
\arguments{
\item{x}{A \link{matrix}.}
}
\value{
Logical indicating whether the given matrix is lower triangular.
}
\description{
Determine if a matrix is upper triangular.
}
\examples{
m <- mat("1, -1, -1, -1; 0, 1, -2, -2; 0, 0, 1, -3; 0, 0, 0, 1")
is.triu(m)
is.triu(eye(3, 5))
}
