% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{ones}
\alias{ones}
\alias{zeros}
\title{Matrix/Array of Ones or Zeros}
\usage{
ones(nrow = 1, ncol = 1, ...)

zeros(nrow = 1, ncol = 1, ...)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{...}{Further dimensions of the array.}
}
\value{
A matrix or array of ones or zeros.
}
\description{
Construct a matrix or multi-way array of all ones or zeros.
}
\examples{
ones(3)  # column matrix of ones
fill(1, 3)
zeros(3, 5)  # 3-by-5 matrix of zeros
fill(0, 3, 5)
}
\seealso{
\code{\link{fill}}, \code{\link{falses}}, \code{\link{trues}}.
}

