% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Pad a Matrix}
\usage{
pad(x, padding = c("constant", "edge", "mean"), pad_width = 1,
  constant_values = 0, end_values, reflect_type, stat_depth, na.rm = FALSE)
}
\arguments{
\item{x}{An object that typically inherits from class \code{"matrix"}.}

\item{padding}{Character string specifying the type of padding. Default is
\code{"constant"}.}

\item{pad_width}{Integer specifying the amount of padding. Default is 
\code{1}.}

\item{constant_values}{Numeric constant to be used when 
\code{padding = "constant"}.}

\item{end_values}{Currently ignored.}

\item{reflect_type}{Currently ignored.}

\item{stat_depth}{Number of values at edge of each axis used to calculate the 
statistic value. Note that \code{stat_depth = 1} is equivalent to 
\code{padding = "edge"}.}

\item{na.rm}{Logical indicating whether or not to remove \code{NA} values
before computing any statistics. Default is \code{FALSE}.}
}
\description{
Pads a matrix, data frame, or other matrix-like object.
}
\examples{
m <- mat('1, 2; 3, 4')
pad(m, padding = "constant", constant_values = 0, pad_width = 2)
pad(m, padding = "edge", pad_width = 2)
image(pad(randn(10, 10), constant_values = NA, pad_width = 5))
}

