% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ramify.R
\docType{package}
\name{ramify}
\alias{ramify}
\alias{ramify-package}
\title{ramify: additional matrix functionality}
\description{
Additional matrix functionality for R. Includes a new wrapper for
the built-in matrix function that extends its functionality by allowing
matrices to be created from character strings and lists. A number of
convenience functions have also been added for users more familiar with
MATLAB or Julia.
}
\details{
The main functions in this package are:
  \itemize{
    \item \code{mat} --- matrix wrapper function
    \item \code{ones} --- creates a matrix of all ones
    \item \code{zeros} --- creates a matrix of all zeros
    \item \code{flatten} --- flatten (i.e., collapse) a matrix to one dimension
    \item \code{resize} --- resize a given matrix
  }
}

