% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{size}
\alias{size}
\title{Dimensions of a Matrix/Array}
\usage{
size(x)
}
\arguments{
\item{x}{A \link{matrix}, \link{array}, or \link{data.frame}.}
}
\value{
The dimensions of the object.
}
\description{
Retrieve the dimensions of a matrix or array.
}
\examples{
m <- mat("1, 3, 5; 7, 9, 11")
size(m)
}
\seealso{
\link{dim}.
}
