% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{tril}
\alias{tril}
\title{Extract Lower Triangular Matrix}
\usage{
tril(x, k = 0, diag = TRUE)
}
\arguments{
\item{x}{A \link{matrix}.}

\item{k}{The sub-diagonal at and below which the matrix is filled.
\code{k = 0} is the main diagonal, while \code{k} < 0 is below it, and  \code{k} > 0 is
above. The default is 0.}

\item{diag}{Logical indicating whether to include the diagonal. Default is
\code{TRUE}.}
}
\description{
Extract the lower triangular part of a matrix.
}
\examples{
tril(ones(5, 5))
tril(ones(5, 5), diag = TRUE)
}
