% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{vcat}
\alias{vcat}
\title{Concatenate Matrices}
\usage{
vcat(...)
}
\arguments{
\item{...}{Arguments to be formed into a list.}
}
\description{
Concatenate along the first (i.e., row) dimension.
}

