% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{argmax}
\alias{argmax}
\alias{argmin}
\title{Row/Column Max/Min Indices}
\usage{
argmax(x, rows = TRUE)

argmin(x, rows = TRUE)
}
\arguments{
\item{x}{A matrix.}

\item{rows}{If \code{TRUE} (the default) the indices of each row max/min is
returned.}
}
\value{
A vector of indices.
}
\description{
Returns the indices of the maximum or minimum values along an axis.
}
\examples{
m <- mat("94, 20, 44; 40, 92, 51; 27, 69, 74")
argmax(m)
argmin(m)
}

