% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\name{atleast_2d}
\alias{atleast_2d}
\title{View Input as an Array with at Least Two Dimensions.}
\usage{
atleast_2d(x)
}
\arguments{
\item{x}{An R object, for example a vector, matrix, array, or data frame.}
}
\value{
The same object, but with a \code{"dim"} attribute.
}
\description{
Ensure that the input has at least two dimensions.
}
\examples{
x <- 1:10
x
atleast_2d(x)
}

