% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mat.R
\name{dmat}
\alias{dmat}
\title{Data Frames}
\usage{
dmat(x, ...)
}
\arguments{
\item{x}{A data vector, character string, or a list.}

\item{...}{Aditional optional arguments passed on to \code{mat}.}
}
\value{
A \code{dataframe}.
}
\description{
Like \code{mat}, \coded{mat} creates a data frame from the given set of
values. These values can be represented by a data vector, a character,
string, or a list of vectors.
}

