% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dmat.R
\name{dmat}
\alias{dmat}
\title{Data Frames}
\usage{
dmat(x, ...)
}
\arguments{
\item{x}{A data vector, character string, or a list.}

\item{...}{Aditional optional arguments passed on to \code{mat}.}
}
\value{
A \code{dataframe}.
}
\description{
Like \code{mat}, \code{dmat} creates a data frame from the given set of
values. These values can be represented by a data vector, a character
string, or a list of vectors.
}
\examples{
dmat('1e-01, 2+5, 3, 4, 5; 6, 7, 8, 9^2, pi', rows = FALSE)
z <- list(a = 1:10, b = 11:20, c = 21:30)
dmat(z)  # list elements form rows
dmat(z, rows= FALSE)  # list elements form columns
}
\seealso{
\code{\link{mat}}, \code{\link{bmat}}.
}

