% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{eye}
\alias{eye}
\title{Identity Matrix}
\usage{
eye(nrow = 1, ncol = nrow)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}
}
\description{
Creates an \code{nrow}-by-\code{ncol} identity matrix.
}

