% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{falses}
\alias{falses}
\title{Matrix of Logical Values}
\usage{
falses(nrow = 1, ncol = 1)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}
}
\description{
Creates an \code{nrow}-by-\code{ncol} matrix of \code{FALSE}s.
}

