% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{fill}
\alias{fill}
\title{Fill a Matrix}
\usage{
fill(x, nrow = 1, ncol = 1)
}
\arguments{
\item{x}{The (single) value to fill the matrix with.}

\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}
}
\description{
Create a matrix filled with the value \code{x}.
}
\examples{
fill(pi, 3, 5)
mat(pi, 3, 5)  # same as 'matrix(pi, 3, 5)'
pi * ones(3, 5)
}
\seealso{
\code{\link{ones}}, \code{\link{zeros}}, \code{\link{falses}}, \code{\link{trues}}, \code{\link{mat}}, \code{\link{matrix}}.
}

