% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{flatten}
\alias{flatten}
\title{Flatten Matrices}
\usage{
flatten(x)
}
\arguments{
\item{x}{A matrix object.}
}
\value{
A numeric vector.
}
\description{
Flatten (i.e., collapse) a matrix to one dimension.
}

