% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{hcat}
\alias{hcat}
\title{Concatenate Matrices}
\usage{
hcat(...)
}
\arguments{
\item{...}{Arguments to be formed into a list.}
}
\description{
Concatenate along the second (i.e., column) dimension.
}

