% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{inv}
\alias{inv}
\title{Matrix Inverse}
\usage{
inv(x, ...)
}
\arguments{
\item{x}{A square numeric or complex matrix}

\item{...}{Additional optional arguments.}
}
\description{
Calculates the inverse of a square matrix.
}
\examples{
m <- 3 * eye(5)
inv(m)
}
\seealso{
\code{\link{solve}}.
}

