% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{linspace}
\alias{linspace}
\title{linearly-spaced Elements}
\usage{
linspace(a, b, n = 50)
}
\arguments{
\item{a}{The starting value of the sequence.}

\item{b}{The final value of the sequence.}

\item{n}{The number of samples to generate. Default is 50.}
}
\description{
Construct a vector of \code{n} linearly-spaced elements from \code{a}
to \code{b}.
}

