% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{linspace}
\alias{linspace}
\title{linearly-spaced Elements}
\usage{
linspace(.start, .stop, n = 100)
}
\arguments{
\item{.start}{}

\item{.stop}{}

\item{n}{}
}
\description{
Construct a vector of \code{n} linearly-spaced elements from \code{.start}
to \code{.stop}.
}

