% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{logspace}
\alias{logspace}
\title{Logarithmically-spaced Elements}
\usage{
logspace(.start, .stop, n = 100)
}
\arguments{
\item{.start}{}

\item{.stop}{}

\item{n}{}
}
\description{
Construct a vector of \code{n} logarithmically-spaced elements from
10^\code{.start} to 10^\code{.stop}.
}

