% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{meshgrid}
\alias{meshgrid}
\title{Rectangular 2-D Grid}
\usage{
meshgrid(x, y = x)
}
\arguments{
\item{x}{Numeric vector representing the first coordinate of the grid.}

\item{y}{Numeric vector representing the second coordinate of the grid.}
}
\value{
a list of matrices, each having class \code{c("matrix", "mat")}.
}
\description{
Creates matrices for vectorized evaluations of 2-D scalar/vector fields over
2-D grids.
}
\examples{
x <- seq(-5, 5, by = 0.1)
y <- seq(-5, 5, by = 0.1)
mg <- meshgrid(x, y)
z1 <- sin(mg[[1]]^2 + mg[[2]]^2) / (mg[[1]]^2 + mg[[2]]^2)
z2 <- mg[[1]] * exp(-mg[[1]]^2 - mg[[2]]^2)
contour(x, y, z1)
filled.contour(x, y, z2, xlim = c(-2, 2), ylim = c(-2, 2))
}
\seealso{
\code{\link{expand.grid}}, \code{\link{contour}}.
}

