% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{ones}
\alias{ones}
\title{Matrix of Ones}
\usage{
ones(nrow, ncol = 1)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}
}
\value{
A matrixof dimension \code{nrow}-by-\code{ncol} filled with ones.
}
\description{
Creates a matrix of all ones.
}

