% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pprint.R
\name{pprint}
\alias{pprint}
\alias{pprint.matrix}
\title{Pretty Printing}
\usage{
pprint(x, ...)

\method{pprint}{matrix}(x, dot.row = 4, dot.col = 4, digits = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"matrix"} or \code{"data.frame"}.}

\item{dot.row}{Abc.}

\item{dot.col}{Def.}

\item{digits}{The minimum number of significant digits to be printed in
values.}
}
\description{
Prettier printing for matrices and data frames.
}
\examples{
pprint(randn(100, 100))
}

