% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pprint.R
\name{pprint}
\alias{pprint}
\alias{pprint.matrix}
\title{Pretty Printing}
\usage{
pprint(x, ...)

\method{pprint}{matrix}(x, rowdots = NULL, coldots = NULL, digits = NULL,
  ...)
}
\arguments{
\item{x}{An object of class \code{"matrix"} or \code{"data.frame"}.}

\item{...}{Additional optional arguments. None are used at present.}

\item{rowdots}{Abc.}

\item{coldots}{Def.}

\item{digits}{The minimum number of significant digits to be printed in
values.}
}
\description{
Prettier printing for matrices and data frames.
}
\examples{
pprint(randn(100, 100))
}

