% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mat.R
\name{print.mat}
\alias{print.mat}
\title{Print Matrices, New-Style}
\usage{
\method{print}{mat}(x, pretty.print = getOption("mat.pretty.print"),
  dot.row = getOption("mat.dot.row"), dot.col = getOption("mat.dot.col"),
  digits, ...)
}
\arguments{
\item{x}{A matrix.}

\item{pretty.print}{Logical indicating whether to print a "prettier" version
of the matrix.}

\item{dot.row}{Integer indicating which row to replace with \code{...}.}

\item{dot.col}{Integer indicating which column to replace with \code{...}.}

\item{digits}{Minimal number of significant digits.}

\item{...}{Additional optional arguments passed onto \code{prmatrix}.}
}
\description{
A new method for printing matrices.
}
\examples{
m <- randn(100, 100)
m
print(m, dot.row = 5, dot.col = 5, digits = 2)
}

