% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramify.R
\docType{package}
\name{ramify-package}
\alias{ramify}
\alias{ramify-package}
\title{ramify: additional matrix functionality}
\description{
Additional matrix functionality for R including: (1) wrappers
for the base matrix function that allows matrices to be created from character
strings and lists (the former is especially useful for creating block
matrices), (ii) better printing of large matrices via a new generic function
for "pretty" printing, and (iii) a number of convenience functions for users
more familiar with other scientific languages like 'Julia', 'Matlab'/'Octave',
or 'Python'+'NumPy'.
}
\details{
To learn more about ramify, read the introductory vignette:
\code{browseVignettes(package = "ramify")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgreenwell/ramify}
  \item Report bugs at \url{https://github.com/bgreenwell/ramify/issues}
}

}
\author{
\strong{Maintainer}: Brandon M. Greenwell \email{greenwell.brandon@gmail.com} (\href{https://orcid.org/0000-0002-8120-0084}{ORCID})

}
\keyword{internal}
