% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{rand}
\alias{rand}
\alias{randn}
\title{Matrix/Array of Random Numbers}
\usage{
rand(nrow = 1, ncol = 1, ..., min = 0, max = 1)

randn(nrow = 1, ncol = 1, ..., mean = 0, sd = 1)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{...}{Further dimensions of the array.}

\item{min/max}{Lower and upper limits of the distribution. Must be finite. (
\code{rand} only).}

\item{mean/sd}{Parameters of the normal distribution. (\code{randn} only).}
}
\value{
A  matrix or array of pseudorandom numbers.
}
\description{
Construct a matrix or multi-way array of random deviates.
}
\details{
\code{rand} generates a matrix or array of uniform random deviates
  while \code{randn} generates a matrix or array of normal random deviates.
}
\examples{
rand(2, 3)  # 2-by-3 matrix of uniform random numbers
rand(2, 3, min = 100, max = 200)
randn(2, 3)  # 2-by-3 matrix of standard normal random variates
randn(2, 3, mean = 10, sd = 0.1)
}
\seealso{
\code{\link{runif}}, \code{\link{rnorm}}.
}

