% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{randi}
\alias{randi}
\title{Matrix/Array of Uniform Random Integers}
\usage{
randi(imax, nrow, ncol = 1, ...)
}
\arguments{
\item{imax}{A positive integer.}

\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{...}{Further dimensions of the array.}
}
\value{
A matrix or array of pseudorandom numbers. If the result is a matrix,
        it will have class \code{c("matrix", "mat")}.
}
\description{
Construct a matrix or multi-way array of uniform random integers.
}
\examples{
randi(2, 5, 5)
}
\seealso{
\code{\link{rand}}, \code{\link{randn}}, \code{\link{sample}}.
}

