% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{randn}
\alias{randn}
\title{Normally Distributed Random Numbers}
\usage{
randn(nrow = 1, ncol = 1, ...)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{...}{Additional optional arguments to be passed on to \code{matrix}.}
}
\description{
Creates an \code{nrow}-by-\code{ncol} matrix of normally distributed random
numbers.
}

