% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{resize}
\alias{resize}
\title{Resize Matrix}
\usage{
resize(x, nrow, ncol, byrow = TRUE)
}
\arguments{
\item{x}{A \code{"matrix"} or \code{R} object.}

\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{byrow}{Logical. If FALSE (the default) the matrix is filled by
columns, otherwise the matrix is filled by rows.}
}
\value{
A matrix of dimension \code{nrow}-by-\code{ncol}.
}
\description{
Returns a new matrix of dimension nrow by ncol using the elements of x.
}

