% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convenience.R
\name{size}
\alias{size}
\title{Dimensions of a Matrix}
\usage{
size(x)
}
\arguments{
\item{x}{A matrix or data frame.}
}
\value{
The dimensions of a matrix.
}
\description{
Retrieve the dimensions of a matrix.
}
\seealso{
\code{\link{dim}}.

m <- mat("1, 3, 5; 7, 9, 11")
size(m)
}

