% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{tr}
\alias{tr}
\title{Trace of a Matrix}
\usage{
tr(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
The sum of the diagonal elements of \code{x}.
}
\description{
Sum of diagonal elements of a matrix.
}
\examples{
tr(ones(5, 10))
x <- replicate(1000, tr(rand(25, 25)))
hist(x)
}

