% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\name{tri}
\alias{tri}
\title{Lower/Upper Triangular Matrix}
\usage{
tri(nrow, ncol = nrow, k = 0, diag = TRUE)
}
\arguments{
\item{nrow}{The desired number of rows.}

\item{ncol}{The desired number of columns.}

\item{k}{The sub-diagonal at and below which the matrix is filled.
\code{k = 0} is the main diagonal, while k < 0 is below it, and
k > 0 is above. The default is 0.}

\item{diag}{Logical indicating whether to include the diagonal. Default is
\code{TRUE}.}
}
\description{
Construct a matrix with ones at and below the given diagonal and zeros
elsewhere.
}
\examples{
tri(5, 5)
tri(5, 5, 2)
tri(5, 5, -1)
}

