% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\name{triu}
\alias{triu}
\title{Extract Upper Triangular Matrix}
\usage{
triu(x, k = 0, diag = TRUE)
}
\arguments{
\item{x}{A matrix.}

\item{k}{Diagonal below which to zero elements. \code{k = 0} (the default) is
the main diagonal, k < 0 is below it and k > 0 is above.}

\item{diag}{Logical indicating whether to include the diagonal. Default is
\code{TRUE}.}
}
\description{
Extract the upper triangular part of a matrix.
}
\examples{
triu(ones(5, 5))
triu(ones(5, 5), diag = FALSE)
}

